Matching optimization and templating finished.  Dynamically typed tests are busted atm.
