+++ OneOf.TemplatingModels/OneOf.TemplatingModels.csproj
@@ -43,9 +43,11 @@
   <ItemGroup>
     <Compile Include="Constants.cs" />
     <Compile Include="Helpers.cs" />
+    <Compile Include="MatchMethodModel.cs" />
     <Compile Include="OneOfModel.cs" />
     <Compile Include="OneOfSwitcherModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SwitchMethodModel.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
